feat: Fix incorrect LM object check in evaluator
The code changes in this commit fix a bug in the evaluator where the check for the LM object was incorrectly performed on the model object instead. This resulted in the LM object not being properly removed when it should have been. The fix ensures that the check is performed on the correct object, lm, and removes the LM object if it is not an instance of LM.

Remove not used txt writer